+++ examples/chat/pages/index.tsx
@@ -6,26 +6,23 @@ import { CardinalTwitterIdentityResolver } from '@dialectlabs/identity-cardinal'
 import { DialectDappsIdentityResolver } from '@dialectlabs/identity-dialect-dapps';
 import { SNSIdentityResolver } from '@dialectlabs/identity-sns';
 import {
-  Backend, ChatButton,
+  Backend,
+  ChatButton,
   Config,
   defaultVariables,
   DialectContextProvider,
   DialectThemeProvider,
-  DialectUiManagementProvider, DialectWalletAdapter, IncomingThemeVariables
+  DialectUiManagementProvider,
+  DialectWalletAdapter,
+  IncomingThemeVariables,
 } from '@dialectlabs/react-ui';
 import {
   useConnection,
   useWallet,
-  WalletContextState
+  WalletContextState,
 } from '@solana/wallet-adapter-react';
 import Head from 'next/head';
 import { Wallet as WalletButton } from '../components/Wallet';
-// pink: #B852DC
-// teal: #59C29D
-// dark: #353535
-// light: #F6F6F6
-// border-light: #F0F0F0
-// blue: #448EF7
 
 // TODO: Use useTheme instead of explicitly importing defaultVariables
 export const themeVariables: IncomingThemeVariables = {
@@ -37,19 +34,19 @@ export const themeVariables: IncomingThemeVariables = {
   },
   light: {
     bellButton:
-      'w-12 h-12 shadow-md hover:shadow-lg shadow-neutral-300 hover:shadow-neutral-400 text-[#59C29D]',
+      'w-12 h-12 shadow-md hover:shadow-lg shadow-neutral-300 hover:shadow-neutral-400 text-teal',
     modal:
-      'sm:border border-[#F0F0F0] shadow-lg shadow-neutral-300 sm:rounded-xl',
+      'sm:border border-border-ligh shadow-lg shadow-neutral-300 sm:rounded-xl',
     colors: {
-      primary: 'text-[#353535]',
+      textPrimary: 'text-dark',
     },
-    button: `${defaultVariables.light.button} border-none bg-[#B852DC]`,
-    highlighted: `${defaultVariables.light.highlighted} bg-[#F6F6F6] border border-[#F0F0F0]`,
-    input: `${defaultVariables.light.input} border-b-[#59C29D] focus:ring-[#59C29D] text-[#59C29D]`,
-    iconButton: `${defaultVariables.light.iconButton} hover:text-[#59C29D] hover:opacity-100`,
-    avatar: `${defaultVariables.light.avatar} bg-[#F6F6F6]`,
-    messageBubble: `${defaultVariables.light.messageBubble} border-none bg-[#448EF7] text-white`,
-    sendButton: `${defaultVariables.light.sendButton} bg-[#59C29D]`,
+    button: `${defaultVariables.light.button} border-none bg-pink`,
+    highlighted: `${defaultVariables.light.highlighted} bg-light border border-border-light`,
+    input: `${defaultVariables.light.input} border-b-teal focus:ring-teal text-teal`,
+    iconButton: `${defaultVariables.light.iconButton} hover:text-teal hover:opacity-100`,
+    avatar: `${defaultVariables.light.avatar} bg-light`,
+    messageBubble: `${defaultVariables.light.messageBubble} border-none bg-blue text-black`,
+    sendButton: `${defaultVariables.light.sendButton} bg-teal`,
   },
 };